backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6

# Data & Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Database (SQLite)
sqlalchemy>=2.0.36
alembic>=1.14.0

# Security
validators>=0.34.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# AI Integration
google-generativeai>=0.8.3

# Tool Integration
python-nmap>=0.7.1
xmltodict>=0.14.0
requests>=2.32.0

# HTTP Client (for webhooks)
httpx>=0.27.0

# System Monitoring
psutil>=6.1.0

# Utilities
loguru>=0.7.2

# SSE Streaming
sse-starlette>=2.0.0